refactor: merge rx::bidi::CrtpBase into rx::CrtpBase
