+++ include/dcc/address.hpp
@@ -52,8 +52,6 @@ struct Address {
   bool reversed{};  /// Direction reversed
 };
 
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Warray-bounds"
 /// Decode address
 ///
 /// \tparam InputIt std::input_iterator
@@ -85,7 +83,6 @@ constexpr Address decode_address(InputIt first) {
   // 255
   else return {*first, Address::IdleSystem};
 }
-#pragma GCC diagnostic pop
 
 /// Encode address
 ///
