+++ tests/rx/app_dyn.cpp
@@ -1,13 +1,16 @@
-#include "bidi_test.hpp"
+#include "rx_test.hpp"
 
-TEST_F(BiDiTest, app_dyn) {
-  _addrs.received = _addrs.primary;
+using namespace dcc::bidi;
+
+TEST_F(RxTest, app_dyn) {
+  // Send whatever packet to get last received address to match primary
+  Receive(dcc::make_function_group_f4_f0_packet(_addrs.primary, 10u));
 
   EXPECT_LT(DCC_RX_BIDI_DEQUE_SIZE, 42uz);
 
   // Add more datagrams than would fit the queue
   for (auto i{0u}; i < 42u; ++i)
-    datagram(DirectionStatusByte{static_cast<uint8_t>(i)});
+    _mock.datagram(DirectionStatusByte{static_cast<uint8_t>(i)});
 
   auto i{0u};
 
@@ -18,8 +21,8 @@ TEST_F(BiDiTest, app_dyn) {
     std::vector<uint8_t> datagram;
     std::ranges::copy(first, std::back_inserter(datagram));
     std::ranges::copy(second, std::back_inserter(datagram));
-    EXPECT_CALL(*this, transmitBiDi(DatagramMatcher(datagram))).Times(1);
-    cutoutChannel2();
+    EXPECT_CALL(_mock, transmitBiDi(DatagramMatcher(datagram))).Times(1);
+    _mock.cutoutChannel2();
   }
 
   {
@@ -28,8 +31,8 @@ TEST_F(BiDiTest, app_dyn) {
     std::vector<uint8_t> datagram;
     std::ranges::copy(third, std::back_inserter(datagram));
     std::ranges::copy(fourth, std::back_inserter(datagram));
-    EXPECT_CALL(*this, transmitBiDi(DatagramMatcher(datagram))).Times(1);
-    cutoutChannel2();
+    EXPECT_CALL(_mock, transmitBiDi(DatagramMatcher(datagram))).Times(1);
+    _mock.cutoutChannel2();
   }
 
   {
@@ -38,8 +41,8 @@ TEST_F(BiDiTest, app_dyn) {
     std::vector<uint8_t> datagram;
     std::ranges::copy(fifth, std::back_inserter(datagram));
     std::ranges::copy(sixt, std::back_inserter(datagram));
-    EXPECT_CALL(*this, transmitBiDi(DatagramMatcher(datagram))).Times(1);
-    cutoutChannel2();
+    EXPECT_CALL(_mock, transmitBiDi(DatagramMatcher(datagram))).Times(1);
+    _mock.cutoutChannel2();
   }
 
   {
@@ -49,13 +52,13 @@ TEST_F(BiDiTest, app_dyn) {
     EXPECT_EQ(DCC_RX_BIDI_DEQUE_SIZE, 7uz);
     std::vector<uint8_t> datagram;
     std::ranges::copy(seventh, std::back_inserter(datagram));
-    EXPECT_CALL(*this, transmitBiDi(DatagramMatcher(datagram))).Times(1);
-    cutoutChannel2();
+    EXPECT_CALL(_mock, transmitBiDi(DatagramMatcher(datagram))).Times(1);
+    _mock.cutoutChannel2();
   }
 
   {
     // Deque is empty again at this point
-    EXPECT_CALL(*this, transmitBiDi(_)).Times(0);
-    cutoutChannel2();
+    EXPECT_CALL(_mock, transmitBiDi(_)).Times(0);
+    _mock.cutoutChannel2();
   }
 }
