+++ tests/rx/backoff_test.cpp
@@ -1,16 +1,16 @@
-#include "bidi_backoff_test.hpp"
+#include "backoff_test.hpp"
 #include <random>
 
-BiDiBackoffTest::BiDiBackoffTest() {
+BackoffTest::BackoffTest() {
   std::random_device rd;
   std::mt19937 gen(rd());
   std::uniform_int_distribution<unsigned int> dist(0, RAND_MAX);
   srand(dist(gen));
 }
 
-BiDiBackoffTest::~BiDiBackoffTest() {}
+BackoffTest::~BackoffTest() {}
 
-int BiDiBackoffTest::CountTillFalse(int max) {
+int BackoffTest::CountTillFalse(int max) {
   int i{};
   for (; i < max; ++i)
     if (!_backoff) break;
