+++ tests/rx/decoder_lock.cpp
@@ -17,7 +17,7 @@ TEST_F(RxTest, cv15_zero_deactivates_decoder_lock) {
   SetUp();
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
-  Expectation write_cv{EXPECT_CALL(_mock, writeCv(cv_addr, cv_value))};
+  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));
   ReceiveAndExecuteTwoIdenticalCvWritePackets(cv_addr, cv_value);
 }
 
@@ -27,6 +27,6 @@ TEST_F(RxTest, cv16_zero_deactivates_decoder_lock) {
   SetUp();
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
-  Expectation write_cv{EXPECT_CALL(_mock, writeCv(cv_addr, cv_value))};
+  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));
   ReceiveAndExecuteTwoIdenticalCvWritePackets(cv_addr, cv_value);
 }
