+++ tests/rx/rx_test.cpp
@@ -2,8 +2,8 @@
 #include <algorithm>
 
 RxTest::RxTest() {
-  _cvs[29uz - 1uz] = 0b1010u;       // Decoder configuration
-  _cvs[1uz - 1uz] = 3u;             // Primary address
+  _cvs[29uz - 1uz] = 0b1010u;  // Decoder configuration
+  _cvs[1uz - 1uz] = static_cast<uint8_t>(_addrs.primary);  // Primary address
   _cvs[19uz - 1uz] = 0u;            // Consist address low byte
   _cvs[20uz - 1uz] = 0u;            // Consist address high byte
   _cvs[15uz - 1uz] = 0u;            // Lock
@@ -17,29 +17,29 @@ RxTest::RxTest() {
   _cvs[65298uz - 1uz] = static_cast<uint8_t>(_cid >> 0u);  // CID low byte
   _cvs[65299uz - 1uz] = _session_id;                       // Session ID
   _cvs[65300uz - 1uz] = static_cast<uint8_t>(
-    0b1100'0000u | _logon_addr >> 8u);  // Logon address high byte
+    0b1100'0000u | _addrs.logon >> 8u);  // Logon address high byte
   _cvs[65301uz - 1uz] =
-    static_cast<uint8_t>(_logon_addr >> 0u);  // Logon address low byte
+    static_cast<uint8_t>(_addrs.logon >> 0u);  // Logon address low byte
 }
 
 RxTest::~RxTest() {}
 
 void RxTest::SetUp() {
-  Expectation read_cv{EXPECT_CALL(_mock, readCv(_))
-                        .WillOnce(Return(_cvs[29uz - 1uz]))
-                        .WillOnce(Return(_cvs[1uz - 1uz]))
-                        .WillOnce(Return(_cvs[19uz - 1uz]))
-                        .WillOnce(Return(_cvs[20uz - 1uz]))
-                        .WillOnce(Return(_cvs[15uz - 1uz]))
-                        .WillOnce(Return(_cvs[16uz - 1uz]))
-                        .WillOnce(Return(_cvs[28uz - 1uz]))
-                        .WillOnce(Return(_cvs[250uz - 1uz]))
-                        .WillOnce(Return(_cvs[251uz - 1uz]))
-                        .WillOnce(Return(_cvs[252uz - 1uz]))
-                        .WillOnce(Return(_cvs[253uz - 1uz]))
-                        .WillOnce(Return(_cvs[65297uz - 1uz]))
-                        .WillOnce(Return(_cvs[65298uz - 1uz]))
-                        .WillOnce(Return(_cvs[65299uz - 1uz]))};
+  EXPECT_CALL(_mock, readCv(_))
+    .WillOnce(Return(_cvs[29uz - 1uz]))
+    .WillOnce(Return(_cvs[1uz - 1uz]))
+    .WillOnce(Return(_cvs[19uz - 1uz]))
+    .WillOnce(Return(_cvs[20uz - 1uz]))
+    .WillOnce(Return(_cvs[15uz - 1uz]))
+    .WillOnce(Return(_cvs[16uz - 1uz]))
+    .WillOnce(Return(_cvs[28uz - 1uz]))
+    .WillOnce(Return(_cvs[250uz - 1uz]))
+    .WillOnce(Return(_cvs[251uz - 1uz]))
+    .WillOnce(Return(_cvs[252uz - 1uz]))
+    .WillOnce(Return(_cvs[253uz - 1uz]))
+    .WillOnce(Return(_cvs[65297uz - 1uz]))
+    .WillOnce(Return(_cvs[65298uz - 1uz]))
+    .WillOnce(Return(_cvs[65299uz - 1uz]));
   _mock.init();
 }
 
@@ -61,16 +61,27 @@ void RxTest::Cutout() {
 }
 
 void RxTest::EnterServiceMode() {
-  Expectation service_mode_hook{EXPECT_CALL(_mock, serviceModeHook(true))};
+  EXPECT_CALL(_mock, serviceModeHook(true));
   Receive(dcc::make_reset_packet());
   Execute();
 }
 
 void RxTest::ReceiveAndExecuteTwoIdenticalCvWritePackets(uint32_t addr,
                                                          uint8_t byte) {
-  auto packet{dcc::make_cv_access_long_write_packet(3u, addr, byte)};
+  auto packet{
+    dcc::make_cv_access_long_write_packet(_addrs.primary, addr, byte)};
   for (auto i{0uz}; i < 2uz; ++i) {
     Receive(packet);
     Execute();
   }
+}
+
+void RxTest::Logon() {
+  EXPECT_CALL(_mock, readCv(_))
+    .WillOnce(Return(_cvs[65300uz - 1uz]))
+    .WillOnce(Return(_cvs[65301uz - 1uz]));
+
+  // Enable
+  Receive(
+    dcc::make_logon_enable_packet(dcc::AddressGroup::Now, _cid, _session_id));
 }
