+++ tests/rx/verify_byte.cpp
@@ -2,9 +2,9 @@
 
 TEST_F(RxTest, verify_byte_operations_mode) {
   auto cv_addr{RandomInterval(0u, smath::pow(2u, 10u) - 1u)};
-  auto packet{dcc::make_cv_access_long_verify_packet(3u, cv_addr)};
+  auto packet{dcc::make_cv_access_long_verify_packet(_addrs.primary, cv_addr)};
 
-  Expectation read_cv{EXPECT_CALL(_mock, readCv(cv_addr, 0u))};
+  EXPECT_CALL(_mock, readCv(cv_addr, 0u));
   Receive(packet);
   Execute();
 }
@@ -16,9 +16,8 @@ TEST_F(RxTest, verify_byte_service_mode) {
   auto packet{dcc::make_cv_access_long_verify_service_packet(cv_addr, 42u)};
 
   // 5 or more identical packets
-  Expectation read_cv{
-    EXPECT_CALL(_mock, readCv(cv_addr, 42u)).WillOnce(Return(42u))};
-  Expectation ack{EXPECT_CALL(_mock, serviceAck())};
+  EXPECT_CALL(_mock, readCv(cv_addr, 42u)).WillOnce(Return(42u));
+  EXPECT_CALL(_mock, serviceAck());
   for (auto i{0uz}; i < 5uz; ++i) {
     Receive(packet);
     Execute();
