+++ tests/rx/write_bit.cpp
@@ -5,8 +5,8 @@ TEST_F(RxTest, write_bit_operations_mode) {
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto bit{RandomInterval(0u, 1u)};
   auto position{RandomInterval(0u, 7u)};
-  auto packet{
-    dcc::make_cv_access_long_write_packet(3u, cv_addr, bit, position)};
+  auto packet{dcc::make_cv_access_long_write_packet(
+    _addrs.primary, cv_addr, bit, position)};
 
   // 2 or more identical packets
   Expectation write_cv{EXPECT_CALL(_mock, writeCv(cv_addr, bit, position))};
